requirements.txt
anytree==2.12.1
contourpy==1.2.1
cycler==0.12.1
dynamixel-sdk==3.7.31
fonttools==4.53.1
graphviz==0.20.3
ikpy==3.4.2
kiwisolver==1.4.5
matplotlib==3.9.2
mpmath==1.3.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pygame==2.6.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.14.1
six==1.16.0
sympy==1.13.2
tabulate==0.9.0
tzdata==2024.1
